+++ Tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -1,0 +1,62 @@
+<?php
+
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
+namespace Symfony\Cmf\Bundle\SimpleCmsBundle\Tests\DependencyInjection;
+
+use Symfony\Cmf\Bundle\SimpleCmsBundle\DependencyInjection\CmfSimpleCmsExtension;
+use Symfony\Cmf\Bundle\SimpleCmsBundle\DependencyInjection\Configuration;
+use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionConfigurationTestCase;
+
+class ConfigurationTest extends AbstractExtensionConfigurationTestCase
+{
+    protected function getContainerExtension()
+    {
+        return new CmfSimpleCmsExtension();
+    }
+
+    protected function getConfiguration()
+    {
+        return new Configuration();
+    }
+
+    public function testSupportsAllConfigFormats()
+    {
+        $expectedConfiguration = array(
+            'persistence' => array(
+                'phpcr' => array(
+                    'enabled' => true,
+                    'basepath' => '/cms/simple',
+                    'manager_registry' => 'doctrine_phpcr',
+                    'manager_name' => null,
+                    'document_class' => 'Symfony\Cmf\Bundle\SimpleCmsBundle\Doctrine\Phpcr\Page',
+                    'use_sonata_admin' => true,
+                    'sonata_admin' => array(
+                        'sort' => 'asc',
+                    )
+                ),
+            ),
+            'use_menu' => false,
+        );
+
+        $formats = array_map(function ($path) {
+            return __DIR__.'/../../Resources/Fixtures/'.$path;
+        }, array(
+            'config/config.yml',
+            'config/config.xml',
+            'config/config.php',
+        ));
+
+        foreach ($formats as $format) {
+            $this->assertProcessedConfigurationEquals($expectedConfiguration, array($format));
+        }
+    }
+}